+++ test/runtests.jl
@@ -15,6 +15,9 @@ end
 @safetestset "Hamiltonian SINDy                                                               " begin
     include("hamiltonian_tests.jl")
 end
+@safetestset "JuliaGNI Conformance                                                            " begin
+    include("conformance_tests.jl")
+end
 @safetestset "Aqua                                                                            " begin
     include("aqua_tests.jl")
 end